Removed unnecessary casting

+++ temp-tracker.ino
@@ -139,5 +139,5 @@ float readTempInC(int sensorPin, float arduinoVoltage)
 
   // Convert from 10 mv per degree with 500 mV offset
   // to degrees: (voltage - 500mV) times 100
-  return (float)((voltage - 0.5) * 100);
+  return (voltage - 0.5) * 100;
 }
